Enable logging in using an access token (#21)
We do a `client.load_store()` to restore a previous session.

If both token and password are defined, token is preferred, since it wont create a new device for the bot. One or the other needs to be defined.

Requires https://github.com/anoadragon453/nio-template/pull/20 for the configuration change since password and access token must both be optional (but one must be given).